+++ trunk/src/lib/1990/1990_duree.h
@@ -22,11 +22,11 @@
 #include "config.h"
 #include "common_projet.h"
 
-int _1990_duree_norme       (unsigned int type,
-                             Type_Pays    pays)
+uint16_t _1990_duree_norme (uint8_t   type,
+                            Type_Pays pays)
                                        __attribute__((__warn_unused_result__));
-char *_1990_duree_norme_txt (unsigned int type,
-                             Type_Pays    pays)
+char *_1990_duree_norme_txt (uint8_t   type,
+                             Type_Pays pays)
                                        __attribute__((__warn_unused_result__));
 
 #endif
